requirements.txt
streamlit>=1.33.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow-cpu>=2.13.0
Pillow>=10.0.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
protobuf>=3.20.0
